requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pillow==12.1.0
requests==2.31.0
python-multipart==0.0.9
cloudinary==1.37.0